--- src/anedya/core.py
@@ -1,20 +1,0 @@
-# src/core.py
-
-import requests
-import datetime
-
-class Anedya:
-    TIME_FORMAT = '%Y-%m-%d %H:%M:%S'
-
-    def __init__(self, user_id, password, api_key):
-        self.user_id = user_id
-        self.password = password
-        self.api_key = api_key
-
-    def send_data(self, data):
-        # Implement your code to send data to Anedya's cloud service using HTTP.
-        pass
-
-    def get_data(self, device_id):
-        # Implement your code to retrieve data from Anedya's cloud service using HTTP.
-        pass
