+++ src/anedya/data/__init__.py
@@ -1,9 +1,8 @@
 #src/anedya/datatypes/__init__.py
-import json
+from .store import *
+from .float import *
 
-class AnedyaEncoder(json.JSONEncoder):
-    def default(self, obj):
-        if hasattr(obj,'toJSON'):
-            return obj.toJSON()
-        else:
-            return json.JSONEncoder.default(self, obj)
+__all__ = [
+    "batch",
+    "float",
+]
