+++ src/anedya/data/encoder.py
@@ -1,0 +1,8 @@
+import json
+
+class AnedyaEncoder(json.JSONEncoder):
+    def default(self, obj):
+        if hasattr(obj,'toJSON'):
+            return obj.toJSON()
+        else:
+            return json.JSONEncoder.default(self, obj)
