+++ src/anedya_dev_sdk_python.egg-info/SOURCES.txt
@@ -1,6 +1,14 @@
 README.md
 setup.cfg
 setup.py
+src/anedya/__init__.py
+src/anedya/anedya.py
+src/anedya/config/__init__.py
+src/anedya/config/config.py
+src/anedya/data/__init__.py
+src/anedya/data/batch.py
+src/anedya/data/encoder.py
+src/anedya/data/float.py
 src/anedya_dev_sdk_python.egg-info/PKG-INFO
 src/anedya_dev_sdk_python.egg-info/SOURCES.txt
 src/anedya_dev_sdk_python.egg-info/dependency_links.txt
