+++ src/anedya_dev_sdk_python.egg-info/top_level.txt
@@ -1,1 +1,1 @@
-
+anedya
